server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
python-dotenv
httpx
pydantic
pydantic-settings
PyJWT
cryptography

# AI & LLM
zhipuai
sniffio
scikit-learn
pandas
numpy

# Document AI (Wave 4)
# pytesseract
# Pillow
# PyPDF2
